PYTHON/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
requests
numpy
scikit-learn
torch
torchvision
Pillow
firebase-admin
asyncio
textwrap3
arxiv
google-adk
google-genai
duckduckgo-search
pypdf
pydantic